feat(workflow): Add metric alert chart to chartcuterie
